LibWeb: Tokenize CSS numbers as doubles
Every later stage uses doubles, so dropping that precision right at the
start of parsing is a little silly. :^)

+++ Userland/Libraries/LibWeb/CSS/Parser/Tokenizer.h
@@ -87,7 +87,7 @@ private:
     [[nodiscard]] ErrorOr<Token> consume_a_numeric_token();
     [[nodiscard]] ErrorOr<Token> consume_an_ident_like_token();
     [[nodiscard]] Number consume_a_number();
-    [[nodiscard]] float convert_a_string_to_a_number(StringView);
+    [[nodiscard]] double convert_a_string_to_a_number(StringView);
     [[nodiscard]] ErrorOr<FlyString> consume_an_ident_sequence();
     [[nodiscard]] u32 consume_escaped_code_point();
     [[nodiscard]] ErrorOr<Token> consume_a_url_token();
